Transfered main to new TestConcatention class, and packed classes
