+++ concatenation/ConcatenateMatrices.java
@@ -1,3 +1,5 @@
+package concatenation;
+
 import java.util.*;
 import java.io.*;
 
@@ -148,40 +150,5 @@ public class ConcatenateMatrices {
                 output.printf("[%d,%d]: %s\n", i, j, array[i][j]);
             }
         }
-    }
-    public static void main(String[] args) throws IOException
-    {
-        Scanner scanner = new Scanner(new File("vstup.txt"));
-        PrintStream output = new PrintStream("vystup.txt");
-
-
-        int rowsNum=0,colsNum=0;
-
-        if(scanner.hasNextLine())
-        {
-            Scanner firstLineScanner = new Scanner(scanner.nextLine());
-            if (firstLineScanner.hasNextInt())
-                rowsNum = firstLineScanner.nextInt();
-            if (firstLineScanner.hasNextInt())
-                colsNum = firstLineScanner.nextInt();
-        }
-
-        List<String[][]> allMatrices = new ArrayList<String[][]>();
-
-        while(scanner.hasNext())
-        {
-            String[][] curMatrix = readNextMatrix(scanner, rowsNum, colsNum);
-
-            if(curMatrix == null)
-                break;
-
-            allMatrices.add(curMatrix);
-        }
-
-        String[][] resultMatrix = concatenatePlentyMatrices(allMatrices);
-
-        arrayToOutput(output, resultMatrix);
-
-        scanner.close();
     }
 }
